+++ src/selectors/TaskTemplateSelectors.js
@@ -1,4 +1,5 @@
 import { createSelector } from 'reselect';
+import { getSettings } from 'selectors/SettingSelectors';
 import { filterByVisibleState } from 'utils/CategoryUtils';
 import { compareStrings } from 'utils/CompareUtils';
 
@@ -16,5 +17,13 @@ export const getVisibleTaskTemplateSelector = () => createSelector(
     (state, id) => id,
     (taskTemplates, id) => {
         return taskTemplates.find(taskTemplate => taskTemplate.id === id);
+    }
+);
+
+export const getDefaultTaskTemplate = createSelector(
+    getTaskTemplatesFilteredByVisibleState,
+    getSettings,
+    (taskTemplates, settings) => {
+        return taskTemplates.find(taskTemplate => taskTemplate.id === settings.defaultTaskTemplate);
     }
 );