+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add work log tab
 * Add status bar settings
 * Add "Add sub-task" action
+* Add shortcuts in date picker
 * Add custom protocol handler
 * Improved batch add and edit tasks dialog
 * Show total elapsed time for today in status bar (#116)
@@ -15,6 +16,7 @@
 * Fix display bugs in note and task tables (#50) (#52)
 * Fix date format in fields
 * Fix Toodledo task completion date
+* Fix usage of default task template
 
 # 0.9.1